third_party/WebKit/LayoutTests/platform/linux/external/wpt/web-animations/timing-model/timelines/document-timelines-expected.txt

This is a testharness.js-based test.
FAIL Document timelines report current time relative to navigationStart assert_equals: The current time matches requestAnimationFrame time expected 30.258999999887237 but got 40.81733333327975
Harness: the test ran to completion.
